[FIX] Structured array coercion strikes again
Was using the wrong dtype for the first download of the PANAS scores for
the Mirchi et al., 2018 example. When you first fetch the data the
returned Y matrix is int, but when you re-load a cached version it was
being returned as float. Should be resolved, now!

+++ netneurotools/datasets/mirchi.py
@@ -178,6 +178,6 @@ def fetch_mirchi2018(data_dir=None, resume=True, verbose=1):
         Y = np.array([tuple(row) for row in np.column_stack(list(Y.values()))],
                      dtype=dict(names=list(Y.keys()), formats=['i8'] * len(Y)))
     else:
-        Y = np.genfromtxt(Y_fname, delimiter=',', names=True)
+        Y = np.genfromtxt(Y_fname, delimiter=',', names=True, dtype=int)
 
     return X, Y
